Att. Sistema de Buffs Permanentes
- agora é mostrado os itens na UI

+++ Assets/Scripts/Save_System/InfosSave.cs
@@ -6,14 +6,14 @@ using UnityEngine;
 [System.Serializable]
 public class InfosSave
 {
-    public float[] quantidadeItem;
+    public int[] quantidadeItem;
 
     //habilidades permanentes
     //informações da narrativa
 
     public InfosSave()
     {
-        quantidadeItem = new float[4];
+        quantidadeItem = new int[4];
         for(int i = 0; i < quantidadeItem.Length; i++)
         {
             quantidadeItem[i] = 0;
